data-app-with-hybrid-tables/requirements.txt
flask
Flask-Caching
snowflake-snowpark-python
python-dotenv
